Rename hres to content
